refactor(upload): 重构上传队列为统一管理架构
重构上传队列系统，引入BaseUploadQueue基类实现通用队列管理逻辑，创建UploadQueueHelper统一管理所有上传队列。将DlassUploadQueue和WebDavUploadQueue重构为继承自BaseUploadQueue的具体实现，简化代码并提高可维护性。修改MainWindow初始化代码以使用新的统一初始化方法。
