+++ Ink Canvas/Helpers/UploadHelper.cs
@@ -36,6 +36,16 @@ namespace Ink_Canvas.Helpers
     /// </summary>
     public class DlassUploadProvider : IUploadProvider
     {
+        private static readonly DlassUploadQueue _queue = new DlassUploadQueue();
+
+        /// <summary>
+        /// 静态构造函数
+        /// </summary>
+        static DlassUploadProvider()
+        {
+            UploadQueueHelper.RegisterQueue(_queue);
+        }
+
         /// <summary>
         /// 提供者名称
         /// </summary>
@@ -54,7 +64,7 @@ namespace Ink_Canvas.Helpers
         /// <returns>是否上传成功</returns>
         public async Task<bool> UploadAsync(string filePath, CancellationToken cancellationToken = default)
         {
-            return await DlassNoteUploader.UploadNoteFileAsync(filePath, cancellationToken);
+            return await _queue.UploadFileAsync(filePath, cancellationToken);
         }
     }
 
@@ -63,6 +73,16 @@ namespace Ink_Canvas.Helpers
     /// </summary>
     public class WebDavUploadProvider : IUploadProvider
     {
+        private static readonly WebDavUploadQueue _queue = new WebDavUploadQueue();
+
+        /// <summary>
+        /// 静态构造函数
+        /// </summary>
+        static WebDavUploadProvider()
+        {
+            UploadQueueHelper.RegisterQueue(_queue);
+        }
+
         /// <summary>
         /// 提供者名称
         /// </summary>
@@ -81,7 +101,7 @@ namespace Ink_Canvas.Helpers
         /// <returns>是否上传成功</returns>
         public async Task<bool> UploadAsync(string filePath, CancellationToken cancellationToken = default)
         {
-            return await WebDavUploadQueue.UploadFileAsync(filePath, cancellationToken);
+            return await _queue.UploadFileAsync(filePath, cancellationToken);
         }
     }
 
@@ -109,6 +129,9 @@ namespace Ink_Canvas.Helpers
                 // 注册默认上传提供者
                 RegisterProviderInternal(new DlassUploadProvider());
                 RegisterProviderInternal(new WebDavUploadProvider());
+
+                // 初始化所有上传队列
+                UploadQueueHelper.InitializeAllQueues();
 
                 _initialized = true;
             }
